Revert "imx8mpevk/igh: enable ethercat native driver ec_fec on imx8mpevk"
This reverts commit 79545377137ab0cc0a2938acb434e40111951de9.

When "ec_fec" native driver is enabled, the original ENET fec driver
must be compiled to moudles. Because modules are loaded later than
build-in drivers, the interface names of ENET and ENET_QOS are exchanged.
It means that the ENET will be renamed "eth1" and the ENET_QOS will be
renamed "eth0". This change will break some existing scripts which use
“eth1” for ENET_QOS.

Due to the above reason, the ec_fec native dirver is not enabled by
default for i.mx8mpevk platform in this release.

If the ec_fec native driver is required, please follow below steps to
enable it. But note that the ENET inteface will be named "eth1".
In the meantime, the ENET_QOS inteface will be renamed "eth0" from "eth1".

 * Configure the fec driver to modules by adding the below line
   on "conf/distro/include/real-time-edge-base.inc" file.

	DELTA_KERNEL_DEFCONFIG_append_mx8mp = " linux-fec.config"
 * Enable the ec_fec native driver by adding the below line
   on "conf/distro/include/igh-ethercat.inc"
	IGH_ETHERCAT_imx8mpevk = " fec "

Signed-off-by: Wenbin Song <wenbin.song@nxp.com>