package: lldpd: add preemption tlv and update link aggregation tlv
Additional Ethernet capabilities TLV are defined in IEEE Std 802.3-2018,
it indicate preemption capability on a port. lldpd has not included this
feature, this patch adds a patch to support it.

Link Aggregation TLV has been deprecated from  IEEE Std 802.3 and is now
standardized as IEEE Std 802.1AX, which is part of the IEEE 802.1. This
patch adds a patch to update it to align the new definition.

Signed-off-by: Xiaoliang Yang <xiaoliang.yang_1@nxp.com>